requirements.txt
python-telegram-bot>=20.0
qrcode>=7.4.2
Pillow>=10.0.0 